Refactor postprocessors
